API reference

Functions and macros

HAML.Codegen.@haml_str — Macro

@haml_str(source)
haml"..."

Include HAML source code into Julia source. The code will be executed in the context (module / function) where it appears and has access to the same variables.

Example

julia> haml"%p Hello, world"
"<p>Hello, world</p>"

HAML.Templates.includehaml — Function

includehaml(mod::Module, fn::Symbol, path, indent="")
includehaml(mod::Module, fns::Pair{Symbol}...)

Define methods for the function mod.fn that allow rendering the HAML template in the file path. These methods have the following signatures:

fn(io::IO; variables...)
fn(f::Function; variables...)
fn(; variables...)

where the output of the template will be written to io / passed to f / returned respectively.

HAML.Templates.render — Function

render(io, path; variables=(), indent="")

Evaluate HAML code in the file specified by path and write the result to io. Any variables passed as variables will be available to the resulting code as $key.

Types

HAML.Parse.Source — Type

HAML.Source("/path/to/file.hamljl")
HAML.Source(::LineNumberNode, ::AbstractString)

Represent Julia-flavoured HAML source code that can be parsed using the Meta.parse function.